refactor: move `getParamsByQuerystring` and `getQuerystringByParams` to `Querystring`
