+++ src/index.ts
@@ -1,4 +1,3 @@
-export * from './getParamsByQuerystring';
-export * from './getQuerystringByParams';
 export * from './getResponseStatus';
 export * from './isNodeJS';
+export * from './Querystring';
